+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/jersey/provider/JerseyProviderInstaller.java
@@ -11,6 +11,7 @@ import org.glassfish.jersey.internal.inject.AbstractBinder;
 import org.glassfish.jersey.internal.inject.InjectionResolver;
 import org.glassfish.jersey.server.monitoring.ApplicationEventListener;
 import org.glassfish.jersey.server.spi.internal.ValueParamProvider;
+import ru.vyarus.dropwizard.guice.module.installer.InstallersOptions;
 import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.AbstractJerseyInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.JerseyManaged;
 import ru.vyarus.dropwizard.guice.module.installer.install.binding.BindingInstaller;
@@ -32,6 +33,19 @@ import static ru.vyarus.dropwizard.guice.module.installer.util.JerseyBinding.*;
 /**
  * Jersey provider installer.
  * Looks for classes annotated with {@link javax.ws.rs.ext.Provider} and register bindings in HK context.
+ * <p>
+ * By default, user providers are prioritized (with {@link org.glassfish.jersey.internal.inject.Custom}
+ * qualifier). This is the default dropwizard behaviour for direct provider registration with
+ * {@code environment.jersey().register(provider)} and so installer behaves the same. Without it ambiguous situations
+ * are possible when dropwizard default providers used instead (e.g. user provided {@code ExceptionMapper<Throwable>}
+ * not used at all because of dropwizard's one).
+ * Auto qualification may be disabled with
+ * {@link ru.vyarus.dropwizard.guice.module.installer.InstallersOptions#PrioritizeJerseyExtensions} (to mimic older
+ * guicey versions behaviour). When auto prioritization disabled, {@link org.glassfish.jersey.internal.inject.Custom}
+ * annotation may be used directly (to prioritize exact providers).
+ * <p>
+ * {@link javax.annotation.Priority} may be used to order providers (see {@link javax.ws.rs.Priorities} for
+ * the default priority constants).
  * <p>
  * If provider is annotated with {@link JerseyManaged} it's instance will be created by HK2, not guice.
  * This is important when extensions directly depends on HK beans (no way to wrap with {@link Provider}
@@ -111,6 +125,7 @@ public class JerseyProviderInstaller extends AbstractJerseyInstaller<Object> imp
     public void install(final AbstractBinder binder, final Injector injector, final Class<Object> type) {
         final boolean hkExtension = isJerseyExtension(type);
         final boolean forceSingleton = isForceSingleton(type, hkExtension);
+        final boolean prioritize = option(InstallersOptions.PrioritizeJerseyExtensions);
         // since jersey 2.26 internal hk Factory class replaced by java 8 Supplier
         if (is(type, Supplier.class)) {
             // register factory directly (without wrapping)
@@ -122,7 +137,7 @@ public class JerseyProviderInstaller extends AbstractJerseyInstaller<Object> imp
                     GenericsResolver.resolve(type).getGenericsInfo().getComposingTypes());
             if (!extensions.isEmpty()) {
                 for (Class<?> ext : extensions) {
-                    bindSpecificComponent(binder, injector, type, ext, hkExtension, forceSingleton);
+                    bindSpecificComponent(binder, injector, type, ext, hkExtension, forceSingleton, prioritize);
                 }
             } else {
                 // no known extension found